requirements.txt
Complete output from command git rev-parse HEAD:
fatal: ambiguous argument 'HEAD': unknown revision or path not in the working tree.
Use '--' to separate paths from revisions, like this:
'git <command> [<revision>...] -- [<file>...]'
HEAD

----------------------------------------
adjustText==0.7.3
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==3.0.2
boto3==1.9.66
botocore==1.12.66
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
configparser==3.5.0
## !! Could not determine repository location
congressideology==0.1.0
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
databricks-cli==0.8.2
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
dnspython==1.15.0
docutils==0.14
entrypoints==0.2.3
eventlet==0.24.1
fasttext==0.8.22
Flask==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
google-api-core==1.6.0
google-auth==1.6.1
google-auth-oauthlib==0.2.0
google-cloud-bigquery==1.8.0
google-cloud-core==0.29.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
greenlet==0.4.15
gunicorn==19.9.0
httplib2==0.12.0
idna==2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterthemes==0.20.0
kiwisolver==1.0.1
lesscpy==0.13.0
lxml==4.2.5
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mleap==0.8.1
mlflow==0.8.0
monotonic==1.5
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.4
nose==1.3.7
nose-exclude==0.5.0
notebook==5.7.2
numpy==1.15.4
oauthlib==2.1.0
pandas==0.23.4
pandas-gbq==0.8.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
plac==0.9.6
ply==3.11
preshed==2.0.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pybind11==2.2.4
Pygments==2.3.0
pyparsing==2.3.0
PySocks==1.6.8
python-congress==0.3.5
python-dateutil==2.7.5
python-dotenv==0.9.1
pytz==2018.7
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
querystring-parser==1.2.3
-e git+https://github.com/crazyfrogspb/RedditScore.git@8d7812c41e83fbd1d8ee9140a2e6484651e99b4c#egg=redditscore
regex==2018.1.10
requests==2.20.1
requests-file==1.4.3
requests-oauthlib==1.0.0
rsa==4.0
s3transfer==0.1.13
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
smmap2==2.0.5
spacy==2.0.18
tabulate==0.8.2
terminado==0.8.1
testpath==0.4.2
thinc==6.12.1
tldextract==2.2.0
toolz==0.9.0
torch==1.0.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
tweepy==3.6.0
ujson==1.35
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
yapf==0.25.0
